Add parquet-filter and sort benchmarks to dfbench
