+++ benchmarks/README.md
@@ -229,31 +229,16 @@ This will produce output like
 └──────────────┴──────────────┴──────────────┴───────────────┘
 ```
 
-### Expected output
-
-The result of query 1 should produce the following output when executed against the SF=1 dataset.
-
-```
-+--------------+--------------+----------+--------------------+--------------------+--------------------+--------------------+--------------------+----------------------+-------------+
-| l_returnflag | l_linestatus | sum_qty  | sum_base_price     | sum_disc_price     | sum_charge         | avg_qty            | avg_price          | avg_disc             | count_order |
-+--------------+--------------+----------+--------------------+--------------------+--------------------+--------------------+--------------------+----------------------+-------------+
-| A            | F            | 37734107 | 56586554400.73001  | 53758257134.870026 | 55909065222.82768  | 25.522005853257337 | 38273.12973462168  | 0.049985295838396455 | 1478493     |
-| N            | F            | 991417   | 1487504710.3799996 | 1413082168.0541    | 1469649223.1943746 | 25.516471920522985 | 38284.467760848296 | 0.05009342667421622  | 38854       |
-| N            | O            | 74476023 | 111701708529.50996 | 106118209986.10472 | 110367023144.56622 | 25.502229680934594 | 38249.1238377803   | 0.049996589476752576 | 2920373     |
-| R            | F            | 37719753 | 56568041380.90001  | 53741292684.60399  | 55889619119.83194  | 25.50579361269077  | 38250.854626099666 | 0.05000940583012587  | 1478870     |
-+--------------+--------------+----------+--------------------+--------------------+--------------------+--------------------+--------------------+----------------------+-------------+
-Query 1 iteration 0 took 1956.1 ms
-Query 1 avg time: 1956.11 ms
-```
-
-# Benchmark Descriptions
+# Benchmark Runner
 
 ## `dfbench`
 
-The `dfbench` program contains subcommands to run various benchmarks.
+The `dfbench` program contains subcommands to run the various
+benchmarks. When benchmarking, it should always be built in release
+mode using `--release`.
 
-Full help can be found in the relevant sub command. For example to get help for tpch,
-run `cargo run --release  --bin dfbench tpch --help`
+Full help for each benchmark can be found in the relevant sub
+command. For example to get help for tpch, run
 
 ```shell
 cargo run --release --bin dfbench  --help
@@ -270,6 +255,22 @@ SUBCOMMANDS:
     tpch            Run the tpch benchmark.
     tpch-convert    Convert tpch .slt files to .parquet or .csv files
 
+```
+
+For example, to run the parquet filter benchmarks, do:
+
+```shell
+cargo run --release --bin dfbench -- parquet-filter  --path ./data --scale-factor 1.0
+
+Generated test dataset with 1232662234232432 rows
+Executing 'Selective-ish filter' (filter: request_method = Utf8("GET"))
+Using scan options ParquetScanOptions { pushdown_filters: false, reorder_filters: false, enable_page_index: false }
+Iteration 0 returned 20438 rows in 310.860411 ms
+Iteration 1 returned 20438 rows in 292.474873 ms
+Iteration 2 returned 20438 rows in 300.402753 ms
+Using scan options ParquetScanOptions { pushdown_filters: true, reorder_filters: true, enable_page_index: true }
+Iteration 0 returned 20438 rows in 367.220051 ms
+Iteration 1 returned 20438 rows in 354.972939 ms
 ```
 
 ## NYC Taxi Benchmark
@@ -317,50 +318,3 @@ h2o groupby query 1 took 1669 ms
 
 [1]: http://www.tpc.org/tpch/
 [2]: https://www1.nyc.gov/site/tlc/about/tlc-trip-record-data.page
-
-## Parquet benchmarks
-
-This is a set of benchmarks for testing and verifying performance of parquet filtering and sorting.
-The queries are executed on a synthetic dataset generated during the benchmark execution and designed to simulate web server access logs.
-
-To run filter benchmarks, run:
-
-```base
-cargo run --release --bin parquet -- filter  --path ./data --scale-factor 1.0
-```
-
-This will generate the synthetic dataset at `./data/logs.parquet`. The size of the dataset can be controlled through the `size_factor`
-(with the default value of `1.0` generating a ~1GB parquet file).
-
-For each filter we will run the query using different `ParquetScanOption` settings.
-
-Example run:
-
-```
-Running benchmarks with the following options: Opt { debug: false, iterations: 3, partitions: 2, path: "./data", batch_size: 8192, scale_factor: 1.0 }
-Generated test dataset with 10699521 rows
-Executing with filter 'request_method = Utf8("GET")'
-Using scan options ParquetScanOptions { pushdown_filters: false, reorder_predicates: false, enable_page_index: false }
-Iteration 0 returned 10699521 rows in 1303 ms
-Iteration 1 returned 10699521 rows in 1288 ms
-Iteration 2 returned 10699521 rows in 1266 ms
-Using scan options ParquetScanOptions { pushdown_filters: true, reorder_predicates: true, enable_page_index: true }
-Iteration 0 returned 1781686 rows in 1970 ms
-Iteration 1 returned 1781686 rows in 2002 ms
-Iteration 2 returned 1781686 rows in 1988 ms
-Using scan options ParquetScanOptions { pushdown_filters: true, reorder_predicates: false, enable_page_index: true }
-Iteration 0 returned 1781686 rows in 1940 ms
-Iteration 1 returned 1781686 rows in 1986 ms
-Iteration 2 returned 1781686 rows in 1947 ms
-...
-```
-
-Similarly, to run sorting benchmarks, run:
-
-```base
-cargo run --release --bin parquet -- sort  --path ./data --scale-factor 1.0
-```
-
-This proceeds in the same way as the filter benchmarks: each sort expression
-combination will be run using the same set of `ParquetScanOption` as the
-filter benchmarks.
